Now, move the file to models module, and update imports.
