+++ tests/cases/apis/HelloAPI/remote/t_Hello.py
@@ -7,7 +7,7 @@ from unittest import TestCase
 from ogd.common.configs.TestConfig import TestConfig
 from ogd.common.utils.Logger import Logger
 # import locals
-from src.ogd.apis.utils.TestRequest import APIRequest
+from src.ogd.apis.models.APIRequest import APIRequest
 from tests.config.t_config import settings
 
 class t_Hello_remote(TestCase):
